Move Providers to its own namespace
