+++ tests/Providers/MongolidUserProviderTest.php
@@ -1,10 +1,12 @@
 <?php
-namespace MongolidLaravel;
+namespace MongolidLaravel\Providers;
 
 use Illuminate\Contracts\Auth\Authenticatable;
 use Illuminate\Contracts\Hashing\Hasher;
 use Mockery as m;
 use MongoDB\BSON\ObjectID;
+use MongolidLaravel\MongolidModel;
+use MongolidLaravel\TestCase;
 
 class MongolidUserProviderTest extends TestCase
 {
